refactor: 네이밍 Controller -> Servlet으로 변경
